FAB in main activity
