+++ app/src/main/java/arslanali/ru/moneytracker/fab/AddFabBehavior.java
@@ -25,12 +25,17 @@ public class AddFabBehavior extends FloatingActionButton.Behavior {
         super.onNestedScroll(coordinatorLayout, child, target, dxConsumed, dyConsumed, dxUnconsumed, dyUnconsumed);
 
         //child -> Floating Action Button
-        if (dyConsumed > 0) {
-            CoordinatorLayout.LayoutParams layoutParams = (CoordinatorLayout.LayoutParams) child.getLayoutParams();
-            int fab_bottomMargin = layoutParams.bottomMargin;
-            child.animate().translationY(child.getHeight() + fab_bottomMargin).setInterpolator(new LinearInterpolator()).start();
-        } else if (dyConsumed < 0) {
-            child.animate().translationY(0).setInterpolator(new LinearInterpolator()).start();
+//        if (dyConsumed > 0) {
+//            CoordinatorLayout.LayoutParams layoutParams = (CoordinatorLayout.LayoutParams) child.getLayoutParams();
+//            int fab_bottomMargin = layoutParams.bottomMargin;
+//            child.animate().translationY(child.getHeight() + fab_bottomMargin).setInterpolator(new LinearInterpolator()).start();
+//        } else if (dyConsumed < 0) {
+//            child.animate().translationY(0).setInterpolator(new LinearInterpolator()).start();
+//        }
+        if (dyConsumed > 0 && child.getVisibility() == View.VISIBLE) {
+            child.hide();
+        } else if (dyConsumed < 0 && child.getVisibility() != View.VISIBLE) {
+            child.show();
         }
     }
 
@@ -40,6 +45,8 @@ public class AddFabBehavior extends FloatingActionButton.Behavior {
                                        View directTargetChild,
                                        View target,
                                        int nestedScrollAxes) {
-        return nestedScrollAxes == ViewCompat.SCROLL_AXIS_VERTICAL;
+        return nestedScrollAxes == ViewCompat.SCROLL_AXIS_VERTICAL ||
+                super.onStartNestedScroll(coordinatorLayout, child, directTargetChild, target,
+                        nestedScrollAxes);
     }
 }
