+++ app/src/main/res/layout/activity_main.xml
@@ -29,4 +29,15 @@
         android:layout_height="match_parent"
         app:layout_behavior="@string/appbar_scrolling_view_behavior" />
 
+    <android.support.design.widget.FloatingActionButton
+        android:id="@+id/fabAdd"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:layout_gravity="bottom|end"
+        android:layout_margin="@dimen/fab_margin"
+        android:src="@mipmap/ic_plus"
+        app:backgroundTint="@color/colorPrimary"
+        app:layout_behavior="arslanali.ru.moneytracker.fab.AddFabBehavior"
+        app:useCompatPadding="true" />
+
 </android.support.design.widget.CoordinatorLayout>
