+++ app/src/main/res/layout/items.xml
@@ -19,15 +19,4 @@
 
     </android.support.v4.widget.SwipeRefreshLayout>
 
-    <android.support.design.widget.FloatingActionButton
-        android:id="@+id/fabAdd"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_gravity="bottom|end"
-        android:layout_margin="@dimen/fab_margin"
-        android:src="@mipmap/ic_plus"
-        app:backgroundTint="@color/colorPrimary"
-        app:layout_behavior="arslanali.ru.moneytracker.FAB"
-        app:useCompatPadding="true" />
-
 </FrameLayout>
